Move com.hazelcast.partition.membergroup.MemberGroup back to public API to com.hazelcast.spi.partitiongroup package
