+++ hazelcast/src/main/java/com/hazelcast/internal/partition/membergroup/ConfigMemberGroupFactory.java
@@ -20,6 +20,7 @@ import com.hazelcast.config.MemberGroupConfig;
 import com.hazelcast.cluster.Member;
 import com.hazelcast.cluster.impl.MemberImpl;
 import com.hazelcast.internal.util.AddressUtil;
+import com.hazelcast.spi.partitiongroup.MemberGroup;
 
 import java.util.Collection;
 import java.util.HashMap;
