+++ hazelcast/src/main/java/com/hazelcast/internal/partition/membergroup/DefaultMemberGroup.java
@@ -17,6 +17,7 @@
 package com.hazelcast.internal.partition.membergroup;
 
 import com.hazelcast.cluster.Member;
+import com.hazelcast.spi.partitiongroup.MemberGroup;
 
 import java.util.Collection;
 import java.util.HashSet;
