+++ hazelcast/src/main/java/com/hazelcast/internal/partition/membergroup/MemberGroupFactory.java
@@ -18,6 +18,7 @@ package com.hazelcast.internal.partition.membergroup;
 
 import com.hazelcast.config.PartitionGroupConfig;
 import com.hazelcast.cluster.Member;
+import com.hazelcast.spi.partitiongroup.MemberGroup;
 
 import java.util.Collection;
