+++ hazelcast/src/main/java/com/hazelcast/spi/partitiongroup/MemberGroup.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.partition.membergroup;
+package com.hazelcast.spi.partitiongroup;
 
 import com.hazelcast.cluster.Member;
