+++ hazelcast/src/main/java/com/hazelcast/spi/partitiongroup/PartitionGroupStrategy.java
@@ -16,7 +16,6 @@
 
 package com.hazelcast.spi.partitiongroup;
 
-import com.hazelcast.internal.partition.membergroup.MemberGroup;
 import com.hazelcast.spi.discovery.AbstractDiscoveryStrategy;
 
 /**
